// client/src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const Missile = __t.object("Missile", {
  id: __t.u64(),
  ownerId: __t.identity(),
  x: __t.f32(),
  y: __t.f32(),
  heading: __t.f32(),
  speed: __t.f32(),
  targetX: __t.f32(),
  targetY: __t.f32(),
  targetShipId: __t.option(__t.u64()),
});
export type Missile = __Infer<typeof Missile>;

export const PhysicsTimer = __t.object("PhysicsTimer", {
  scheduledId: __t.u64(),
  scheduledAt: __t.scheduleAt(),
});
export type PhysicsTimer = __Infer<typeof PhysicsTimer>;

export const Player = __t.object("Player", {
  identity: __t.identity(),
  nickname: __t.option(__t.string()),
  online: __t.bool(),
});
export type Player = __Infer<typeof Player>;

export const Ship = __t.object("Ship", {
  id: __t.u64(),
  ownerId: __t.identity(),
  get shipClass() {
    return ShipClass;
  },
  x: __t.f32(),
  y: __t.f32(),
  heading: __t.f32(),
  speed: __t.f32(),
  radarOn: __t.bool(),
  get waypoint() {
    return __t.option(Waypoint);
  },
});
export type Ship = __Infer<typeof Ship>;

// The tagged union or sum type for the algebraic type `ShipClass`.
export const ShipClass = __t.enum("ShipClass", {
  ArleighBurke: __t.unit(),
  Carrier: __t.unit(),
});
export type ShipClass = __Infer<typeof ShipClass>;

export const Waypoint = __t.object("Waypoint", {
  x: __t.f32(),
  y: __t.f32(),
  targetSpeed: __t.f32(),
});
export type Waypoint = __Infer<typeof Waypoint>;